Release the HDC when disposing graphics
