+++ src/System.Drawing.Common/tests/GraphicsTests.cs
@@ -673,7 +673,6 @@ namespace System.Drawing.Tests
             }
         }
 
-        [ActiveIssue(20884, TestPlatforms.AnyUnix)]
         [ConditionalFact(Helpers.GdiplusIsAvailable)]
         public void Flush_Disposed_ThrowsArgumentException()
         {
